# api/v1alpha1/openapi.yaml
openapi: 3.0.4
info:
  contact: {}
  description: |
    DCM Environment Agent API — external Service Provider registration and
    management. The agent acts as the intermediary between DCM and Service
    Providers deployed in a target environment.
  title: Environment Agent API
  version: v1alpha1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: /api/v1alpha1

security:
  - bearerAuth: []

tags:
  - name: provider
    description: Service Provider registration and management operations
  - name: health
    description: Health check operations

paths:
  /health:
    get:
      tags:
        - health
      summary: Health check
      operationId: getHealth
      description: Liveness health check for the Environment Agent
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "Unauthorized — missing or invalid JWT Bearer token"
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
  /providers:
    get:
      tags:
        - provider
      summary: List all Service Providers
      operationId: listProviders
      description: |
        Returns all registered Service Providers (both embedded and external)
        including their current health status.
      parameters:
        - name: max_page_size
          in: query
          description: Maximum number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - name: page_token
          in: query
          description: |
            Token for retrieving the next page of results.
            Obtained from the next_page_token field of a previous response.
          schema:
            type: string
      responses:
        '200':
          description: List of registered providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderList'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "Unauthorized — missing or invalid JWT Bearer token"
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - provider
      summary: Register an external Service Provider
      operationId: createProvider
      description: |
        Register an external Service Provider with the agent.
        Registration is idempotent (name as natural key, create-or-update).
        Only one SP — embedded or external — may serve a given service type
        per agent. If the requested service type is already served by another
        SP, registration is rejected with 409 Conflict.
      parameters:
        - name: id
          in: query
          description: Optional provider ID for idempotent registration
          schema:
            type: string
            pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
            minLength: 1
            maxLength: 63
            x-validated-by: handler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: Provider updated (re-registration / lease renewal)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '201':
          description: Provider created (new registration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "Unauthorized — missing or invalid JWT Bearer token"
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            Conflict — service type already served by another provider
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation exception
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

  /providers/{provider_id}:
    get:
      tags:
        - provider
      summary: Get a Service Provider
      operationId: getProvider
      description: |
        Returns a single registered Service Provider by ID, including its
        current health status.
      parameters:
        - name: provider_id
          in: path
          required: true
          description: Unique identifier of the provider
          schema:
            type: string
            pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
            minLength: 1
            maxLength: 63
      responses:
        '200':
          description: Provider found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "Unauthorized — missing or invalid JWT Bearer token"
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Provider not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Provider:
      type: object
      description: Service Provider registration resource
      x-aep-resource:
        type: agent.dcm.io/provider
        singular: provider
        plural: providers
        patterns:
          - providers/{provider_id}
      required:
        - name
        - endpoint
        - service_type
        - schema_version
      properties:
        id:
          type: string
          pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
          minLength: 1
          maxLength: 63
          readOnly: true
          description: Unique identifier for the Service Provider
          example: "sp-db-001"
        path:
          type: string
          readOnly: true
          description: Resource path identifier
          example: "providers/sp-db-001"
        name:
          type: string
          description: Unique name of the Service Provider (natural key)
          example: "db-provider"
        display_name:
          type: string
          description: Human-readable display name for the provider
          example: "Database Service Provider"
        endpoint:
          type: string
          format: uri
          description: URL where the agent can reach the SP
          example: "https://sp.example.com:8080"
        service_type:
          type: string
          description: Type of service this provider offers (one per SP)
          example: "database"
        schema_version:
          type: string
          description: Schema version of the service type the SP supports
          pattern: "^v[0-9]+(alpha|beta)?[0-9]*$"
          example: "v1alpha1"
          x-validated-by: handler
        operations:
          type: array
          items:
            type: string
          description: List of operations supported for this service type
          example: ["create", "delete", "update"]
        metadata:
          $ref: '#/components/schemas/ProviderMetadata'
        type:
          type: string
          readOnly: true
          description: Whether the SP is embedded or external
          enum:
            - embedded
            - external
          example: "external"
        status:
          type: string
          readOnly: true
          description: Current health state of the provider
          enum:
            - Ready
            - Unhealthy
            - Unavailable
          example: "Ready"
        last_check_time:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp of the last health check
          example: "2026-06-05T10:30:00Z"
        create_time:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the provider was first registered
        update_time:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the provider was last updated

    ProviderList:
      type: object
      description: Paginated list of providers
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Provider'
        next_page_token:
          type: string
          description: Token for retrieving the next page of results

    Health:
      type: object
      description: Health status singleton resource
      x-aep-resource:
        singular: health
        plural: healths
        singleton: true
        type: agent.dcm.io/health
        patterns:
          - health
      properties:
        status:
          type: string
          description: Health status
          example: "healthy"
        path:
          type: string
          description: Canonical path of the resource
          example: "health"
          readOnly: true

    ProviderMetadata:
      type: object
      description: Additional metadata about the provider
      additionalProperties: true
      properties:
        region_code:
          type: string
          description: Geographic region code where the provider operates
          example: "us-east-1"
        zone:
          type: string
          description: Availability zone or datacenter identifier
          example: "datacenter-b"
        status:
          type: string
          description: Current operational status of the provider
          example: "healthy"
        resources:
          $ref: '#/components/schemas/ResourceCapacity'

    ResourceCapacity:
      type: object
      description: Resource capacity information
      properties:
        total_cpu:
          type: string
          description: Total CPU cores available
          example: "200"
        total_memory:
          type: string
          description: Total memory available
          example: "1TB"
        total_storage:
          type: string
          description: Total storage available
          example: "2TB"
        total_node:
          type: integer
          description: Total number of nodes
          example: 100

    Error:
      type: object
      description: |
        RFC 9457 compliant error response (Problem Details for HTTP APIs).

        Defined problem types:

          | Type URI | Title | HTTP Status |
          |----------|-------|-------------|
          | .../problems/invalid-argument | Invalid argument | 400 |
          | .../problems/unauthenticated | Unauthenticated | 401 |
          | .../problems/permission-denied | Permission denied | 403 |
          | .../problems/not-found | Not found | 404 |
          | .../problems/already-exists | Already exists | 409 |
          | .../problems/unprocessable-entity | Unprocessable entity | 422 |
          | .../problems/internal | Internal Server Error | 500 |
          | .../problems/unavailable | Service unavailable | 503 |

        Note: permission-denied is defined in the enum but not yet emitted by
        any handler. Titles follow the humanized-slug convention shared with
        the ACM Cluster SP and K8s Container SP.
      required:
        - type
        - title
      properties:
        type:
          type: string
          format: uri-reference
          description: URI reference identifying the error type
          enum:
            - https://dcm-project.github.io/problems/invalid-argument
            - https://dcm-project.github.io/problems/unauthenticated
            - https://dcm-project.github.io/problems/permission-denied
            - https://dcm-project.github.io/problems/not-found
            - https://dcm-project.github.io/problems/already-exists
            - https://dcm-project.github.io/problems/unprocessable-entity
            - https://dcm-project.github.io/problems/internal
            - https://dcm-project.github.io/problems/unavailable
          x-enum-varnames:
            - ErrorTypeINVALIDARGUMENT
            - ErrorTypeUNAUTHENTICATED
            - ErrorTypePERMISSIONDENIED
            - ErrorTypeNOTFOUND
            - ErrorTypeALREADYEXISTS
            - ErrorTypeUNPROCESSABLEENTITY
            - ErrorTypeINTERNAL
            - ErrorTypeUNAVAILABLE
          example: "https://dcm-project.github.io/problems/already-exists"
        title:
          type: string
          description: Short human-readable summary of the problem
          example: "Already exists"
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: HTTP status code
          example: 409
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: "service type 'vm' is already served by provider 'vm-provider'"
        instance:
          type: string
          format: uri-reference
          description: URI reference for this specific error occurrence